+++ src/lib/components/interactables/draggableOptions.interface.ts
@@ -1,9 +1,0 @@
-export interface DraggableOptions {
-    manualStart?: boolean;
-    max?: number;
-    maxPerElement?: number;
-    axis?: 'x' | 'y' | 'xy';
-    autoScroll?: boolean;
-    allowFrom?: string;
-    ignoreFrom?: string;
-}