+++ src/lib/components/interactables/gridOptions.interface.ts
@@ -1,8 +1,0 @@
-export interface GridOptions {
-    x: number;
-    y: number;
-    range?: number;
-    offset?: { x: number; y: number };
-    endOnly?: boolean;
-    relativePoints?: Array<{ x: number; y: number }>;
-}